Correctly obtain HOME directory in Makefile
Thanks to gpt

You can use the following command in a Makefile to get the home directory of the user:
HOME_DIR := $(shell echo $$HOME)

This uses the $(shell ...) syntax to run the echo $$HOME command, which returns
the value of the HOME environment variable, which is set to the home directory
of the user. The output of this command is then assigned to the HOME_DIR
variable.

In a Makefile, the tilde ~ is not a reliable way to refer to the home directory
because the shell that interprets the Makefile may not expand the tilde ~ in the
same way as a shell run from the command line.

Using the $$HOME environment variable, as shown in the previous answer, is a
more portable and reliable method for getting the home directory in a Makefile.
This approach works across different operating systems and shells and will
return the correct value even if the Makefile is run by a different user or in a
different environment.